% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biopsy.R
\name{bio_par}
\alias{bio_par}
\title{Find Banff Score In String}
\usage{
bio_par(fstring, fpara)
}
\arguments{
\item{fstring}{A string to parse}

\item{fpara}{A string. The Banff score ("a", "cg", "mm"...)}
}
\description{
Find Banff Score In String
}
