% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biopsy.R
\name{bio_par_range_to_int}
\alias{bio_par_range_to_int}
\title{Convert Range To Int}
\usage{
bio_par_range_to_int(fstring)
}
\arguments{
\item{fstring}{A string. Interval to change into a integer}
}
\description{
Convert Range To Int
}
