% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biopsy.R
\name{biopsy_parse}
\alias{biopsy_parse}
\title{Parse a string into Banff Classification scores}
\usage{
biopsy_parse(data, col, eliminate_intervals = TRUE, na_to_zeroes = FALSE)
}
\arguments{
\item{data}{A data frame or a data frame extension (a tibble).}

\item{col}{Name of the column containing strings to parse}

\item{eliminate_intervals}{Logical value. When \code{TRUE} transforms score intervals into integers. Default \code{TRUE}.}

\item{na_to_zeroes}{Logical value. When \code{TRUE} transforms \code{NA} values (not found or unable to classify) to \code{0}. Default \code{FALSE}.}
}
\description{
Parse a string (e.g.a unstructured written report) into structured form of Banff Classification scores.
Creates multiple new columns for each classifier.
}
