% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_join_after}
\alias{check_join_after}
\title{Display Status Of The Join  Check Function}
\usage{
check_join_after(
  data,
  ID,
  removed_rows = FALSE,
  group_size = FALSE,
  check_occurance = NA,
  show_removed = FALSE
)
}
\arguments{
\item{data}{A dataframe. Required.}

\item{ID}{Column name of the join by atribute. Required.}

\item{removed_rows}{Logical value. When \code{TRUE} displays removed rows. Default \code{FALSE}.}

\item{group_size}{Logical value. When \code{TRUE} displays table of group sizes by ID. Default \code{FALSE}.}

\item{check_occurance}{Integer. Show rows from group (group_size group) of specific size. Default \code{NA}.}

\item{show_removed}{Logical value. When \code{TRUE} displays removed rows from before the join. Default \code{FALSE}.}
}
\description{
Place this function after joining two or more dataframes
}
