% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_join_before}
\alias{check_join_before}
\title{Initiate The Join Check function}
\usage{
check_join_before(data)
}
\arguments{
\item{data}{A dataframe}
}
\description{
Place this function before joining two or more dataframes
}
