% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labs.R
\name{labs_near_date}
\alias{labs_near_date}
\title{Filter values by the nearest value to a date}
\usage{
labs_near_date(
  data,
  patient_id,
  data_date,
  lab_data,
  lab_data_date,
  lab_data_value,
  matching = "exact",
  omit_na = TRUE,
  duplicit_values = FALSE,
  interval_min = NA,
  interval_max = NA
)
}
\arguments{
\item{data}{A dataframe containing subject IDs and dates to filter by}

\item{patient_id}{A column of unique IDs of a patient or a subject}

\item{data_date}{A column of date values}

\item{lab_data}{A dataframe of laboratory values with Patient's ID, Sample date and Sample Value}

\item{lab_data_date}{A column of Sample dates}

\item{lab_data_value}{A column of Sample Values}

\item{matching}{Controls the matching dates mode. "exact" or "interval"
\itemize{
\item "exact" matches dates (\code{date}, and \code{lab_date}) to be the same
\item "interval" matches \code{lab_date} to be within a defined time interval of \code{date}
}}

\item{omit_na}{Whether to omit missing Sample Values, TRUE by default}

\item{duplicit_values}{Removes duplicit Samples from the same day}

\item{interval_min}{Used when \code{matching} set to "interval". Defines the lower bound of the interval in days.}

\item{interval_max}{Used when \code{matching} set to "interval". Defines the upper bound of the interval in days.}
}
\description{
Match nearest sample values (e. g. laboratory results, measurements) to a date by subject IDs (e.g. patient ID).
}
